(GH-373) Convert pdkNewClass command to feature
This commit converts pdkNewClassCommand to be part of
PdkCommandFeature.
