+++ src/commands/pdkcommands.ts
@@ -2,20 +2,13 @@ import * as vscode from 'vscode';
 import { PDKCommandStrings } from '../messages';
 import { IConnectionManager } from '../connection';
 import { ILogger } from '../logging';
-import { PDKNewClassCommand } from './pdk/pdkNewClassCommand';
 import { PDKNewTaskCommand } from './pdk/pdkNewTaskCommand';
 import { PDKValidateCommand } from './pdk/pdkValidateCommand';
 import { PDKTestUnitCommand } from './pdk/pdkTestCommand';
 
 export function setupPDKCommands(langID: string, connManager: IConnectionManager, ctx: vscode.ExtensionContext, logger: ILogger, terminal: vscode.Terminal) {
-  let newClassCommand = new PDKNewClassCommand(logger, terminal);
-  ctx.subscriptions.push(newClassCommand);
-  ctx.subscriptions.push(vscode.commands.registerCommand(PDKCommandStrings.PdkNewClassCommandId, () => {
-    newClassCommand.run();
-  }));
-  
   let newTaskCommand = new PDKNewTaskCommand(logger, terminal);
-  ctx.subscriptions.push(newClassCommand);
+  ctx.subscriptions.push(newTaskCommand);
   ctx.subscriptions.push(vscode.commands.registerCommand(PDKCommandStrings.PdkNewTaskCommandId, () => {
     newTaskCommand.run();
   }));
